+++ client/src/App.css
@@ -56,6 +56,7 @@
   padding: 0.5rem;
   margin: 0.5rem 0;
   width: 100%;
+  color: black
 }
 
 .login-image {
