+++ client/src/components/Footer/index.jsx
@@ -6,14 +6,14 @@ const Footer = () => {
   return (
     <footer className="w-100 mt-auto text-dark p-4">
       <div className="container text-center mb-5">
-        {location.pathname !== '/' && (
+        {/* {location.pathname !== '/' && (
           <button
             className="btn btn-dark mb-3"
             onClick={() => navigate(-1)}
           >
             &larr; Go Back
           </button>
-        )}
+        )} */}
         <h4>&copy; {new Date().getFullYear()} - Git Fit</h4>
       </div>
     </footer>
