+++ client/src/components/LoginForm/index.jsx
@@ -48,7 +48,7 @@ const LoginForm = (props) => {
             {data ? (
               <p>
                 Success! You may now head{" "}
-                <Link to="/Home">back to the homepage.</Link>
+                <Link to="/">back to the homepage.</Link>
               </p>
             ) : (
               <form onSubmit={handleFormSubmit}>
